feat(cardcontainer): convert to typescript
